+++ ibis/backends/polars/tests/test_udf.py
@@ -4,7 +4,7 @@ import pytest
 
 import ibis.expr.datatypes as dt
 import ibis.expr.types as ir
-from ibis.udf.vectorized import elementwise, reduction
+from ibis.legacy.udf.vectorized import elementwise, reduction
 
 pytest.importorskip("polars")
 pc = pytest.importorskip("pyarrow.compute")
