+++ ibis/interactive.py
@@ -2,7 +2,7 @@ import ibis
 import ibis.examples as ex
 from ibis import deferred as _
 from ibis import selectors as s
-from ibis.expr.operations import udf
+from ibis import udf
 
 ibis.options.interactive = True
