+++ ibis/legacy/udf/validate.py
